+++ controllers/contacts/updateFavoriteStatus.js
@@ -1,4 +1,5 @@
 import Contact from "#models/contactModel.js";
+
 export async function updateFavoriteStatus(req, res, next) {
   const contactId = req.params.contactId;
   const { favorite } = req.body;
