+++ controllers/users/loginUser.js
@@ -1,12 +1,7 @@
 import User from "#models/userModel.js";
-import Joi from "joi";
 import bcrypt from "bcrypt";
+import { loginSchema } from "#validators/loginSchema.js";
 
-// Definiowanie schematu Joi do walidacji danych wejściowych
-const loginSchema = Joi.object({
-  email: Joi.string().email().required(),
-  password: Joi.string().min(6).required(),
-});
 export async function loginUser(req, res) {
   const { email, password } = req.body;
   // Walidacja danych wejściowych
