+++ routes/api/users.js
@@ -1,33 +1,29 @@
+import express from "express";
+
+import verifyToken from "#middleware/authMiddleware.js";
 import { loginUser } from "#controllers/users/loginUser.js";
+import { logoutUser } from "#controllers/users/logoutUser.js";
+import { updateAvatar } from "#controllers/users/updateAvatar.js";
+import { getCurrentUser } from "#controllers/users/getCurrentUser.js";
+import { getExistingUser } from "#controllers/users/getExistingUser.js";
+import { registerNewUser } from "#controllers/users/registerNewUser.js";
 import { loginExistingUser } from "#controllers/users/loginExistingUser.js";
 import { logoutExistingUser } from "#controllers/users/logoutExistingUser.js";
-import { logoutUser } from "#controllers/users/logoutUser.js";
 import { registerExistingUser } from "#controllers/users/registerExistingUser.js";
-import { registerNewUser } from "#controllers/users/registerNewUser.js";
-import { getExistingUser } from "#controllers/users/getExistingUser.js";
-import { getCurrentUser } from "#controllers/users/getCurrentUser.js";
-import verifyToken from "#middleware/authMiddleware.js";
-import { updateAvatar } from "#controllers/users/updateAvatar.js";
-import express from "express";
 
 const router = express.Router();
 
-// Rejestracja użytkownika
 router.post("/signup", registerNewUser);
 router.post("/signup", registerExistingUser);
 
-// Logowanie użytkownika
 router.post("/login", loginUser);
 router.post("/login", loginExistingUser);
 
-// Pobranie danych bieżącego użytkownika
 router.get("/current", verifyToken, getCurrentUser);
 router.get("/current", verifyToken, getExistingUser);
 
-// Wylogowanie użytkownika
 router.get("/logout", verifyToken, logoutExistingUser);
 router.get("/logout", verifyToken, logoutUser);
-// Dodaj nową trasę do aktualizacji avatara
 router.patch("/avatars", verifyToken, updateAvatar);
 
 export default router;
